Finish step 1 of refactoring
Significantly simplify batch training and testing
by uniting model classes and accessing the classes
through a model_factory utility function.

This utility function is a dump if/else switch
given a string return me a model.

So we can build it up as much as we want with
special strings and such.

It does the basic job for now.